Create CMDadmin.bat
Save this code in a file with a .bat extension, for example, run_cmd_as_admin.bat. When you run this batch file, it checks for administrative privileges. If the current user has administrative privileges, it runs CMD as administrator. If not, it relaunches itself with elevated privileges using the User Account Control (UAC) prompt.